id: 5792 fixed non-working close of datepicker in < IE11
ID: 5792
Removed `ELEMENT.focus()` during the picker close.
Problem occured after updating this lib on the 23th of July. Since then
we have an additional option `editable: true` which causes an refocus
durign close. However this does not provide any benefit and the < IE11
is not able to handle it properly, i.e. an new open event is trigger.